+++ CHANGELOG.rst
@@ -24,6 +24,7 @@ Fixed
     - Vue app ontology edit modal search input disabling (from v1.3.0-WIP) (#2454)
     - ``IrodsAccessTicket`` objects deleted on sheet replace (#2460)
     - ``check_igv_file_path()`` partial path matching (#2464)
+    - Germline study app redundant ``get_pedigree_file_path()`` calls (#2465)
 
 
 v1.2.8 (2026-03-19)